move to examples folder with readme
