+++ examples/keyboard/keyboard-key-events.gif
