+++ examples/keyboard/keyboard-pointer-events.gif
