+++ examples/keyboard/use-stdin.ts
